% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LRTval.R
\name{LRTval}
\alias{LRTval}
\title{Testing whether need to use varying coefficience}
\usage{
LRTval(X, Z, Y, p = 1, nknots = 10, df = 1, sigma.output = diag(rep(1,
  length(Y))))
}
\arguments{
\item{X}{predictors, a vector}

\item{Z}{assist variable, a vector}

\item{Y}{dependent variable, a vector}

\item{p}{order of B-splines, p = 1 by default}

\item{df}{the numbers of constraint of fixed coefficient, df = 1 by default, can't be larger than p}

\item{sigma.output}{covariance matrix of disturbance term u}

\item{nkonts}{the number of knots}
}
\description{
Testing whether need to use varying coefficience
}

