% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/valpredict.R
\name{valpredict}
\alias{valpredict}
\title{Prediction for the model}
\usage{
valpredict(model, newx, newz)
}
\arguments{
\item{model}{the object of 'val' generated by val function}

\item{newx}{new predictors, a vector}

\item{newz}{new assist variable, a vector}
}
\description{
Prediction for the model
}

